Feedback: make GC event stashing more pay for play
